requirements.txt
apify-client==1.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-dateutil==2.9.0
